Entire frontend
Entire frontend, with some adjustments to the API for CSRF exemptions,
as well as adding an endpoint to check for the existence of username and
email addresses when registering as a new user. Against better judgement
to push this much in one commit, but needed to get this project out!